+++ src/core_modules/capture-core-utils/featuresSupport/support.js
@@ -1,11 +1,13 @@
 // @flow
 export const FEATURES = Object.freeze({
     programStageWorkingList: 'programStageWorkingList',
+    storeProgramStageWorkingList: 'storeProgramStageWorkingList',
 });
 
 // The first minor version that supports the feature
 const MINOR_VERSION_SUPPORT = Object.freeze({
     [FEATURES.programStageWorkingList]: 39,
+    [FEATURES.storeProgramStageWorkingList]: 40,
 });
 
 export const hasAPISupportForFeature = (minorVersion: string, featureName: string) =>